moved expandable to About section, updated most text in about section
