+++ src/components/children/Headings/About.tsx
@@ -1,33 +1,57 @@
+import Expandable from './reusable/Expandable';
+import { DiagramLanguage } from '../../../interfaces';
+
+const diagrams: DiagramLanguage[] = [
+  {
+    language: 'English',
+    abbreviation: 'EN',
+    text: 'English',
+    href: '/pdf/General Plan Amendment Process.pdf',
+  },
+  {
+    language: 'Spanish',
+    abbreviation: 'SP',
+    text: 'Spanish',
+    href: '/pdf/GPA Process_sp_accessible.pdf',
+  },
+  {
+    language: 'Traditional Chinese',
+    abbreviation: 'ZH',
+    text: 'Chinese - Traditional',
+    href: '/pdf/GPA Process_tc_accessible.pdf',
+  },
+  {
+    language: 'Simplified Chinese',
+    abbreviation: 'ZH',
+    text: 'Chinese - Simplified ',
+    href: '/pdf/GPA Process_sc_accessible.pdf',
+  },
+];
+
 const About = () => {
   return (
     <section tabIndex={0} className="border-t border-gray-300 mb-8 p-8">
       <h2 className="text-3xl yeseva font-bold text-burnt-orange">
-        About the GPA Process
+        About the GPA Process and Outreach
       </h2>
 
       <p className="font-medium pt-4">
-        State Parks has determined that a GPA to the LASHP General Plan is
-        needed to allow for public transit to the LASHP and allow consideration
-        and review of the proposed LA ART public transit use and facility.
+      State Parks determined that an LASHP General Plan amendment is needed to include 
+      a transit land use in the General Plan to allow consideration and review of the proposed aerial rapid transit gondola project. 
       </p>
 
       <p className="font-medium pt-4">
-        State Parks will gather public input and work with LA Aerial Rapid
-        Transit Technologies, LLC (LA ARTT LLC) to identify and address the
-        potential issues and benefits of the project in order to consider a
-        draft GPA proposal in 2025. Once State Parks determines the GPA meets
-        the required standards and conditions for consideration of action, the
-        GPA would be scheduled for a State Park and Recreation Commission (SPRC)
-        meeting agenda.
+        The LASHP GPA public outreach effort began in November 2024 when State Parks held two initial public meetings.
+        State Parks developed General Plan Amendment options based on public feedback and input provided in Fall 2024. 
+        The purpose of this round of public outreach is to present these General Plan Amendment options for public review and feedback.
       </p>
       <p className="font-medium pt-4">
-        State Parks' GPA process will involve a public outreach effort to gather
-        input to help inform the Department's consideration of the LASHP GPA and
-        the LA ART project. There will be multiple opportunities for the public
-        to comment and provide input to the proposed action. California State
-        Parks will conduct several public meetings throughout the GPA process to
-        provide information and receive comment and input.
+        State Parks' General Plan Amendment process involves a public outreach effort to provide information and to gather 
+        input for the State Park and Recreation Commission’s consideration of the LASHP General Plan Amendment. 
+        California State Parks will conduct several public meetings throughout the GPA process to provide information and receive comment and input. 
       </p>
+
+      <Expandable props={diagrams} />
     </section>
   );
 };
